requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
tavily-python>=0.3.0
